% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\name{fitted_marginals}
\alias{fitted_marginals}
\alias{fitted_marginals,portvine_roll-method}
\title{Accessor method for the fitted marginal models of \verb{(cond_)portvine_roll}
objects}
\usage{
fitted_marginals(roll, ...)

\S4method{fitted_marginals}{portvine_roll}(roll)
}
\arguments{
\item{roll}{Object of class \code{portvine_roll} or a child class}

\item{...}{Additional parameters for child class methods}
}
\value{
named list with an entry for each asset containing a
\code{\link[rugarch:ugarchroll-methods]{rugarch::ugarchroll}} class object that encompasses the marginal model fit.
}
\description{
The marginal models are ARMA-GARCH models fitted in a rolling
window fashion using \code{\link[rugarch:ugarchroll-methods]{rugarch::ugarchroll}}.
}
\seealso{
\code{\linkS4class{portvine_roll}}
}
