% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\name{fitted_vines}
\alias{fitted_vines}
\alias{fitted_vines,portvine_roll-method}
\title{Accessor method for the fitted vine copula models of \verb{(cond_)portvine_roll}
objects}
\usage{
fitted_vines(roll, ...)

\S4method{fitted_vines}{portvine_roll}(roll)
}
\arguments{
\item{roll}{Object of class \code{portvine_roll} or a child class}

\item{...}{Additional parameters for child class methods}
}
\value{
list of \code{\link[rvinecopulib:vinecop]{rvinecopulib::vinecop}} class objects each entry
corresponds to one fitted vine copula model for the respective vine window.
}
\description{
Accessor method for the fitted vine copula models of \verb{(cond_)portvine_roll}
objects
}
\seealso{
\code{\linkS4class{portvine_roll}}
}
