% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{marginal_settings-class}
\alias{marginal_settings-class}
\alias{marginal_settings}
\title{S4 class for the marginal settings}
\usage{
marginal_settings(
  train_size,
  refit_size,
  individual_spec = list(),
  default_spec = default_garch_spec()
)
}
\value{
Object of class \code{marginal_settings}
}
\description{
Specify which marginal models (\code{individual_spec} & \code{default_specs}) are
fitted and how often they are refit as well as how big the training data
set is. Remember that the forecasting is done in a rolling window fashion.
}
\details{
For specifying the list for \code{individual_spec} or the argument \code{default_spec}
the function \code{\link[=default_garch_spec]{default_garch_spec()}} might
come in handy.
}
\section{Functions}{
\itemize{
\item \code{marginal_settings}: Class constructor taking the arguments specified below
}}

\section{Slots}{

\describe{
\item{\code{train_size}}{Positive count specifying the training data size.}

\item{\code{refit_size}}{Positive count specifying size of the forecasting
window.}

\item{\code{individual_spec}}{A named list. Specify ARMA-GARCH models for individual
assets by naming the list entry as the asset and providing a
\code{\link[rugarch:ugarchspec-methods]{rugarch::ugarchspec}} object.}

\item{\code{default_spec}}{\code{\link[rugarch:ugarchspec-methods]{rugarch::ugarchspec}}
object specifing the default marginal model (used if model is not specified through
\code{individual_spec})}
}}

\examples{
marginal_settings(100, 10)
}
\seealso{
\code{\link[=default_garch_spec]{default_garch_spec()}}
}
