% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roll_residuals}
\alias{roll_residuals}
\title{Extract fitted residuals from a uGARCHroll object}
\usage{
roll_residuals(ugarchroll, roll_num = 1)
}
\arguments{
\item{ugarchroll}{Object of class \code{\link[rugarch:ugarchroll-methods]{rugarch::ugarchroll}}.}

\item{roll_num}{Count that specifies the fitted model to extract the
residuals from.}
}
\value{
Numeric vector of the fitted residuals.
}
\description{
The \code{\link[rugarch:ugarchroll-methods]{rugarch::ugarchroll}} class object encompasses fitting information
about a number of
models fitted in a rolling window fashion. This utility function gives an
easy interface to extract the fitted residuals from one of these models.
This can be especially helpful for assessing the model quality with a
residual analysis.
}
