% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{sample_returns_small}
\alias{sample_returns_small}
\title{A sample of log returns for 3 assets.}
\format{
data.table with 3 columns and 1000 rows, columns \code{GOOG}, \code{AAPL},
\code{AMZN} contain the daily log return of the 3 stocks.
}
\source{
\href{https://finance.yahoo.com}{Yahoo Finance}
}
\usage{
data(sample_returns_small)
}
\description{
Data extracted from Yahoo Finance representing the daily og returns for
Google, Apple, and Amazon stocks between 2014-01-13 and 2018-01-01 which
results in exactly 1000 observations.
}
\examples{
data(sample_returns_small)
head(sample_returns_small)
}
\keyword{datasets}
