% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_risk_roll.R
\name{estimate_risk_roll}
\alias{estimate_risk_roll}
\title{TBD}
\usage{
estimate_risk_roll(
  data,
  weights = NULL,
  marginal_settings,
  vine_settings,
  alpha = 0.05,
  risk_measures = c("VaR", "ES_mean"),
  cond_vars = NULL,
  n_samples = 1000,
  n_cond_samples = 100,
  n_mc_samples = 1000,
  trace = TRUE
)
}
\arguments{
\item{data}{matrix, data.frame or other object coercible to a data.table
storing the numeric asset returns in the named columns (at least 3).
Moreover missing values must be imputed beforehand.}

\item{weights}{corresponding named non-negative weights of the assets
(conditioning variables must have weight 0).}

\item{marginal_settings}{containing the needed information for the ARMA-GARCH
fitting i.e. marginal models.}

\item{vine_settings}{containing needed information for the vine fitting}

\item{alpha}{a numeric vector specifying the levels in (0,1) at which the
risk measures should be calculated}

\item{risk_measures}{a character vector with valid choices for risk
measures to compute}

\item{cond_vars}{colnames of the variables to sample conditionally from
(currently \eqn{\leq 2})}

\item{n_samples}{number of samples to compute for the risk measure estimates}

\item{n_cond_samples}{number of samples of the conditioning variables}

\item{n_mc_samples}{number of samples for the Monte Carlo integration
if the risk measure \code{ES_mc} is used. (See \code{\link[=est_es]{est_es()}})}

\item{trace}{if set to TRUE the algorithm will print information while
running.}
}
\value{
In the unconditional case an S4 object of class \code{portvine_roll} and
in the conditional case its child class \code{cond_portvine_roll}. For details
see \code{\linkS4class{portvine_roll}}.
}
\description{
TBD
}
\examples{
#TBD
}
